megaraid_sas: Add driver workaround for PERC5/1068 kdump kernel panic
Patchwork-id: 41862
O-Subject: [PATCH 6.2] megaraid_sas: Add driver workaround for PERC5/1068 kdump
	kernel panic
Bugzilla: 723218
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

This is for bz#723218

This hack fixes a problem with older PERC cards, they aren't resetted
properly when the second kernel in kdump starts. The old commands then cause
a panic in ~ 25% of cases.
This could be fixed in firmware, but the hw vendor isn't going to do that.
I've tested this http://www.spinics.net/lists/linux-scsi/msg54294.html
Patch below is a backport of a LSI's polished  version of the same patch
http://www.spinics.net/lists/linux-scsi/msg54924.html

The patch touches only the kdump code (reset_devices)
so it should be safe even for 6.2

Brew - taskID=3706081

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>