GFS2: Fix ->page_mkwrite() races
Patchwork-id: 41882
O-Subject: [RHEL 6.2] GFS2: Fix ->page_mkwrite() races (bz #725091)
Bugzilla: 725091
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

This patch is a RHEL6 port of the upstream patch (nearly identical)
to improve the GFS2 ->page_mkwrite() function such that it will
always return a locked page or an error code.

The core code appears to have been updated recently in RHEL6
causing a coherency issue to be spotted by QE in this area,
and this patch resolves the problem (confirmed by QE testing)

The upstream patch has been posted to cluster-devel, but will
not be appearing in the upstream -nmw tree until hera is back
to normal, unfortunately. It is however queued for upstream
inclusion as soon as possible.

This fixes bz #725091

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>