qla4xxx: export address/port of connection
Patchwork-id: 41915
O-Subject: [RHEL6.2 PATCH 1/1] qla4xxx: export address/port of connection (fix
	boot/udev)
Bugzilla: 728156
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 728156.

Udev uses the connection's persistent address/port in the
/dev/disk/by-path name. Anaconda then uses the same info to decide
if to set things up to use the iscsi tools for boot. qla4xxx is
not exporting this value, so the iscsi tools were not getting
put in the initramfs to be run during boot.

In the old driver we did not have this information but
we do now and can fix this by just setting the flags to tell
libiscsi/scsi_transport_iscsi to export this info in sysfs.

Patch has been sent to linux-scsi (spinics/marc have not updated
yet to show it).

I have tested the patch by logging into the target and
verifying the correct info is now being displayed.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>