reduce overhead on paravirt functions
Patchwork-id: 41950
O-Subject: [RHEL6.2 Patch] Prevent 25% performance hit when running SAP test
	on a kernel with CONFIG_PARAVIRT enabled on hardware.
Bugzilla: 743554
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Dave Anderson <anderson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Chris Wright <chrisw@redhat.com>

We noticed a 25% performance hit running SAP on RHEL6 compared to RHEL5.
After investigating this and profiling the kernel I realized that the
problem only
occurs when CONFIG_PARAVIRT is enabled.

SAP does lots of mprotect system calls so the Profile of a system
running SAP with
CONFIG_PARAVIRT enabled looks like this:

3243.00 53.0% __ptep_modify_prot_start  [kernel]
1390.00 22.7% mprotect_fixup            [kernel]
  620.00 10.1% __ptep_modify_prot_commit [kernel]
  501.00  8.2% native_set_pte_at         [kernel]
   37.00  0.6% native_flush_tlb          [kernel]
   32.00  0.5% intel_idle                [kernel]

Where the profile without CONFIG_PARAVIRT enabled looks like this:

2272.00 48.3% mprotect_fixup
  498.00 10.6% memsetU16
  368.00  7.8% memcpy
  144.00  3.1% intel_idle
  112.00  2.4% bcd_div_dec
   68.00  1.4% prio_tree_replace

The big difference is __ptep_modify_prot_start, __ptep_modify_prot_commit and
native_flush_tlb consume lots of CPU time when CONFIG_PARAVIRT is enabled.

When a kernel with CONFIG_PARAVIRT is enabled the calls through pv_mmu_ops
get patched to static inline functions at boot time with the exception of
.ptep_modify_prot_start, .ptep_modify_prot_commit and .set_pte_at.  The
inline code for these 3 are too big to fit in the calling text so the paravirt
patching does not patch those three.  This results in full subroutine calls
instead of inline code for ptep_modify_prot_start(), ptep_modify_prot_commit()
and set_pte_at() called from mprotect_fixup().  The result is SAP runs 25%
slower on RHEL6 than it ran on RHEL5.

The simplest and cleanest way of eliminating the call overhead and its
associated performance hit when running a kernel built with CONFIG_PARAVIRT
on bare metal is to check if we are running as a paravirt guest and invoke
the inline code directly if not.

The attached patch eliminates the 25% overhead and fixes BZ743554

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>