kabi: Add missing compat_alloc_user_space symbol to kABI
Patchwork-id: 41961
O-Subject: [PATCHv3 RHEL6.2 BZ747047] kabi: Add missing compat_alloc_user_space
	symbol to kABI
Bugzilla: 747047
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 747047
https://bugzilla.redhat.com/show_bug.cgi?id=747047

Description:
============
KABI BZ https://bugzilla.redhat.com/show_bug.cgi?id=703125

Request to add following symbols to kABI:
	compat_alloc_user_space

When updating the megaraid_sas driver for DUP, the symbol
compact_alloc_user_space popped up when verifying kabi. Digging in, it was
noticed that this symbol changed from an inline in 6.0 to a regular function
in 6.1. We would like to add this to the kabi whitelist to make it easier to
maintain future DUPs with this driver.

Looks like this function is not likely to change (as per discussion with Oleg),
and is being used by some modules.

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3717669

Tested:
=======
brewed test kernel properly

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>