tracing: Update the comm field in the right variable in update_max_tr
Patchwork-id: 41963
O-Subject: [PATCH RHEL6.2 BZ736955] tracing: Update the comm field in the right
	variable in update_max_tr
Bugzilla: 736955
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Steven Rostedt <srostedt@redhat.com>

Bugzilla: 736955
https://bugzilla.redhat.com/show_bug.cgi?id=736955

Description:
============
The latency output showed:

The comm is missing in the "task:" and it looks like a minus 3 is
the output. The correct display should be:

The problem is that the comm is being stored in the wrong data
structure. The max_tr.data[cpu] is what stores the comm, not the
tr->data[cpu].

Before this patch the max_tr.data[cpu]->comm was zeroed and the /debug/trace
ended up showing just the '-' sign followed by the pid.

Also remove a needless initialization of max_data.

Upstream status:
================
commit 1acaa1b2d9b5904c9cce06122990a2d71046ce16
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Fri Mar 5 18:23:50 2010 -0300

Brew:
=====
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3722634

Tested:
=======
wakeup tracer shows proper task name

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>