ipc/mqueue: switch back to using non-max values on create
Patchwork-id: 41987
O-Subject: [Patch RHEL6.2 3/6] ipc/mqueue: switch back to using non-max values
	on create
Bugzilla: 746606 746898
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Motohiro Kosaki <mkosaki@redhat.com>

Commit b231cca4381ee15ec99afbfb244fbc0324869927 changed
how we create a queue that does not include an attr
struct passed to open so that it creates the queue
with whatever the maximum values are.  However, if the
admin has set the maximums to allow flexibility in
creating a queue (aka, both a large size and large queue
are allowed, but combined they create a queue too large
for the RLIMIT_MSGQUEUE of the user), then attempts to
create a queue without an attr struct will fail.  Switch
back to using acceptable defaults regardless of what
the maximums are.

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit 35b4c59844f809afbfd9ea039a41cad0ef69a757)

Conflicts:

	ipc/mqueue.c

Note: the git commit ID is from my upstream github tree,
not from Andrew's for-next queue.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>