ipc/mqueue: Up the hard limit on message queues per namespace
Patchwork-id: 41991
O-Subject: [Patch RHEL6.2 6/6] ipc/mqueue: Up the hard limit on message queues
	per namespace
Bugzilla: 746606
RH-Acked-by: Motohiro Kosaki <mkosaki@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

We have a request from a customer for a higher number of messages
queues in a single namespace (bz746606).  After inspecting the code,
a large number of queues does not degrade performance (there is no
silly things like walking a list of queues, where as there is a
walking list of messages in queue on send, so a user would actually
get better performance from the kernel by using lots of queues
with only a few messages than a few queues with lots of messages).
So, increase the maximum number of queues allowed.

Resolves: bz746606

Signed-off-by: Doug Ledford <dledford@redhat.com>

Note: this patch is not yet upstream but will be submitted on the
next merge window.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>