ipc/mqueue: update maximums for the mqueue subsystem
Patchwork-id: 41994
O-Subject: Re: [Patch RHEL6.2 5/6] ipc/mqueue: update maximums for the mqueue
	subsystem
Bugzilla: 746606
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Motohiro Kosaki <mkosaki@redhat.com>

ipc/mqueue: update maximums for the mqueue subsystem

Commit b231cca4381ee15ec99afbfb244fbc0324869927 changed
the maximum size of a message in a message queue from
INT_MAX to 8192*128.  Unfortunately, we had customers
that relied on a size much larger than 8192*128 on their
production systems.  After reviewing POSIX, we found that
it is silent on the maximum message size.  We did find
a couple other areas in which it was not silent.  Fix up
the mqueue maximums so that the customer's system can
continue to work, and document both the POSIX and real
world requirements in ipc_namespace.h so that we don't
have this issue crop back up.

Also, commit 9cf18e1dd74cd0061d58ac55029784ca3dd88f6a
fiddled with HARD_MSGMAX without realizing that the
number was intentionally in place to limit the msg
queue depth to one that was small enough to kmalloc
an array of pointers (hence why we divided 128k by
sizeof(long)).  If we wish to meet POSIX requirements,
we have no choice but to change our allocation to
a vmalloc instead (at least for the large queue size
case).  With that, it's possible to increase our
allowed maximum to the POSIX requirements (or more if
we choose).

Signed-off-by: Doug Ledford <dledford@redhat.com>
(cherry picked from commit b5c368456b64c0eb8e53cdbf50e54a2124ba286b)

Conflicts:

ipc/mqueue.c

Note: the commit ID is from my upstream github tree,
not from Andrew's for-next queue.

Note2: this patch differs from the upstream patch in
Andrew's queue in that patchcheck said I needed
parenthesis around complex expressions in a #define,
and as opposed to doing another commit to add those
as we did in upstream, I rolled that into this patch.
in mqueue.c since we now use vmalloc for large message queues.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>