Add an experimental new driver to support Sandy Bridge CPU's
Patchwork-id: 41999
O-Subject: [PATCH RHEL6 1/3] BZ#647700 edac: Add an experimental new driver to
	support Sandy Bridge CPU's
Bugzilla: 647700
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This is the same patch as submitted upstream:
	http://lkml.org/lkml/2011/10/20/351
and added on my edac tree for kernel 3.2 submission:
	http://git.infradead.org/users/mchehab/edac.git/commit/f828766a26484a3c9285231a40dae7bbb3c57d2d

This driver is known to work on mine and Tony's test environments,
using software error injection, and a partial hardware/software
error injection tool.

There's no broader range test yet to double check if the error decoding
logic will actually point to the right DIMM, so use it with care.
More tests are required to be sure that the driver will work on all
different types of memory configurations.

If you're willing to risk using it, I suggest you to enable EDAC debugs
for your test machines, as the debug logs helps to track what's going
inside the driver.

Please feed me with bug reports, if you notice that the driver
is miss-behaving.

Tested-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>