Fix incorrect edac mode reporting in sb_edac
Patchwork-id: 42000
O-Subject: [PATCH RHEL6 2/3] BZ#647700 EDAC: Fix incorrect edac mode reporting
	in sb_edac
Bugzilla: 647700
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

This is the same patch as submitted upstream:
	https://lkml.org/lkml/2011/10/20/350
And added on my edac git tree for 3.2:
	http://git.infradead.org/users/mchehab/edac.git/commit/ab6bc08a4bc34c0962dbed9f95b996735661d424

From: Mark A. Grondona <mgrondona@llnl.gov>

The edac driver for Sandy Bridge was found to be reporting "FPM"
for edac_mode, which clearly doesn't make sense. It was found that
sb_edac.c:get_dimm_config was reusing a variable for both mem_type
and edac_type, and thus was overwriting the value after setting
it correctly. This patch fixes that issue.

Before the patch:
/sys/devices/system/edac/mc/mc0/csrow0/edac_mode:FPM
/sys/devices/system/edac/mc/mc0/csrow1/edac_mode:FPM
/sys/devices/system/edac/mc/mc0/csrow2/edac_mode:FPM
/sys/devices/system/edac/mc/mc0/csrow3/edac_mode:FPM

After:
/sys/devices/system/edac/mc/mc0/csrow0/edac_mode:S4ECD4ED
/sys/devices/system/edac/mc/mc0/csrow1/edac_mode:S4ECD4ED
/sys/devices/system/edac/mc/mc0/csrow2/edac_mode:S4ECD4ED
/sys/devices/system/edac/mc/mc0/csrow3/edac_mode:S4ECD4ED

Signed-off-by: Mark A. Grondona <mgrondona@llnl.gov>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>