dasd: fix UID readout for z/VM
Patchwork-id: 42008
O-Subject: [RHEL6.2 PATCH 2/3] [s390x]dasd: fix UID readout for z/VM
Bugzilla: 746000

Description
-----------
The UID may not include the extended minidisk token
provided by z/VM and scripts do not recognize the DASD
again.

The buffer used for read configuration data needs to be
initialized with a special EBCDIC string to show support
for extended UIDs to z/VM. This buffer may get
overwritten and re-used for retries and thus does not
contain the needed string.

After a read configuration data returns, a special
function is called that re-initializes the string if
the request was not successful.

Bugzilla
--------
BZ 746000
https://bugzilla.redhat.com/show_bug.cgi?id=746000

Upstream status of the patch
----------------------------
The patch is queued for inclusion into the next kernel release:
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/features

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>