qdio: EQBS retry after CCQ 96
Patchwork-id: 42007
O-Subject: [RHEL6.2 PATCH 3/3] [s390x]qdio: EQBS retry after CCQ 96
Bugzilla: 747578

Description
-----------
A qdio device might no longer respond under z/VM with QIOASSIST on.
This affects qeth- and HiperSockets network devices, as well as
zfcp (SCSI) devices.

If the EQBS instruction returned less than the specified number
of SBALs the queue was not scanned again which could lead to a
stall condition.

Add an instant retry for EQBS with CCQ 96 and if that fails set
up a timer to scan the queue later.

Bugzilla
--------
BZ 747578
https://bugzilla.redhat.com/show_bug.cgi?id=747578

Upstream status of the patch
----------------------------
The patch is queued for inclusion into the next kernel release:
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/features

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>