log userspace: Allow for 'log device name' response in CTR msg exchange
Patchwork-id: 42033
O-Subject: [RHEL6.2 PATCH - v2] DM log userspace: Allow for 'log device name'
	response in CTR msg exchange
Bugzilla: 746254
RH-Acked-by: Alasdair Kergon <agk@redhat.com>

Changes from previous:
1) Unconditional call to kfree
2) Documentation now states device string must be NULL terminated.
3) Code checks for NULL-terminated device string and errors out if the string
   is ill-formed rather than trying to cope with it.

Bug #746254:
 https://bugzilla.redhat.com/show_bug.cgi?id=746254

brew build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3738625

Patch Description:
dm-log-userspace is a module that allows device-mapper mirror logs to be
processed in userspace.  It acts as a communication bridge between the log API
and userspace servers.

Since all log processing belongs to the userspace server, there was no way for
the kernel module to call 'dm_get_device' on the log device.  This call is
essential for properly creating dependency chains that are later queried by
applications like LVM.  Without this correct information, LVM does not function
properly.

The solution is to allow return data from the userspace log server during the
instantiation of the log.  The return data contains the name of the log device
(if there is one) that the kernel can then use to call 'dm_get_device'.  This
solves the problem of the unlinked log.

Testing:
I've tested that this solution is backwards compatible from the userspace and
kernel perspective.  If the kernel and userspace log server have been updated,
the new information is passed down to the kernel; allowing the device to be
registered.  If the kernel is new, but the log server is old; the log server
will not pass down any device information and the kernel will simply bypass the
device registration as before.  If the kernel is old but the log server is new,
the log server will see the old version # and not pass down the device info.
(If it did pass the device name down in error, the kernel would report an error
due to having insufficient space to accept the new information.)

Probing LVM after it queries the kernel for dependencies on a 2-way mirror
should reveal 3 devices - the two mirror images and the log device.  Before
the patch, we see only 2:
(gdb) p *deps
$1 = {count = 2, filler = 0, device = 0x7f673c01f438}
After the patch, we see 3:
(gdb) p *deps
$4 = {count = 3, filler = 0, device = 0x77a9a8}

Upstream status:
Agk has accepted the upstream patch, but it has not yet been pulled.
Corresponding userspace patches are already checked-in upstream.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>