libfc: revert fix deadlock bug in fc_exch_abort_locked
Patchwork-id: 42041
O-Subject: [RHEL6.2 PATCH 1/4] libfc: revert fix deadlock bug in
	fc_exch_abort_locked
Bugzilla: 745667
RH-Acked-by: Andy Grover <agrover@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>

From: Mike Christie <mchristi@redhat.com>

rhel 6 kernel commit:
commit 3849d2202a35a689ff4ea59f5bbaf0c8458c38a1
Author: Mike Christie <mchristi@redhat.com>
Date:   Thu Sep 8 10:44:49 2011 -0400

    [scsi] libfc: fix fc_eh_host_reset

added a a regression we tried to fix with:

commit f7c29d8eea2e853c6ef04c91257ccc5b7777cba6
Author: Mike Christie <mchristi@redhat.com>
Date:   Wed Oct 5 00:33:22 2011 -0400

    [scsi] libfc: fix deadlock bug in fc_exch_abort_locked

but that did not completely fix the problem and still caused this
https://lists.open-fcoe.org/pipermail/devel/2011-October/011787.html.

The following patches fix the first regression and second one.
The patch "libfc: fix deadlock bug in fc_exch_abort_locked" is
no longer needed and so this patch reverts it.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>