lockless queuecommand dispatch fixup
Patchwork-id: 42112
O-Subject: [RHEL6.2 PATCH-V2] lockless queuecommand dispatch fixup.
Bugzilla: 749018
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

bz749018

Remove un-needed check of (scsi-cmnd->serial_number == 0)

This was conflicting with removal of scsi_cmd_get_serial()
in scsi_dispach_cmd when lockless, which previously initialized
scsi-cmnd->serial_number to non-zero.

The use of serial_number=0 checking during error processing
is not required because blk_mark_rq_complete ensures that
the command is either aborted or completed.

This 2nd version doesn't remove the setting of the
scsi-cmnd->serial_number to 0 during error handling
as this more closely mirrors the behavior of rhel6
prior to the introduction of lockless queue command
functionality, for the non-lockless case.

As Tomas Henzl pointed out in reponse to the previous
version of this patch, upstream kept this initialization
as well.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>