mlx4_en: fix endianness with blue frame support
Patchwork-id: 42121
O-Subject: [PATCH RHEL6.2 BZ750166] mlx4_en: fix endianness with blue frame
	support
Bugzilla: 750166
RH-Acked-by: Doug Ledford <dledford@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>

RHBZ#:
------
https://bugzilla.redhat.com/show_bug.cgi?id=750166

Description:
------------
The doorbell register was being unconditionally swapped. In x86, that
meant it was being swapped to BE and written to the descriptor and to
memory, depending on the case of blue frame support or writing to
doorbell register. On PPC, this meant it was being swapped to LE and
then swapped back to BE while writing to the register. But in the blue
frame case, it was being written as LE to the descriptor.

The fix is not to swap doorbell unconditionally, write it to the
register as BE and convert it to BE when writing it to the descriptor.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@linux.vnet.ibm.com>
Reported-by: Richard Hendrickson <richhend@us.ibm.com>
Cc: Eli Cohen <eli@dev.mellanox.co.il>
Cc: Yevgeny Petrilin <yevgenyp@mellanox.co.il>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

Aris: PeterM would like Doug to ack this. this currently doesn't have
all of the acks yet for RHEL 6.2, but posting to start the tech review.

RHEL Version Found:
-------------------
RHEL 6.2

Brew:
-----
http://brewweb.devel.redhat.com/brew/taskinfo?taskID=3756954

Upstream:
---------
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;
a=commit;h=c5d6136e10d667965e46f998c01863802f7b89c8

Test Status:
------------

Thadeu Lima de Souza Cascardo helped me test this.

---------------------------------------------------------------
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>