UV: KABI breakage in uv_blade_info
Patchwork-id: 42141
O-Subject: [RHEL6.2 PATCH v4] UV: KABI breakage in uv_blade_info
Bugzilla: 745253
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

bz745253: UV: KABI breakage in uv_blade_info
Fourth Post

This retracts and partially redoes bz689026 because
of KABI breakage with SGI modules.  We did not catch
this earlier because of a work-around in our build
system to avoid bz737753 during the later stages of
rhel6.2 development.

The second patch put nmi_lock and nmi_count into
a separate array index by blade, thereby better
tracking the upstream patch.  Also, during testing
an infinite NMI loop was uncovered on a large
machine.  The problem is that the time required
to dump all the stack backtraces exceeded the
NMI watchdog.  The delay loop is a hack to avoid
the problem.

The third post removes the commented out
    spin_lock(&uv_nmi_lock)
that was replaced by the
    spin_trylock() and touch_nmi_watchdog()
loop; also removed the change in priority to
be compliant with upstream.

This fourth post adds comments and makes the
uv_blade_info_nmi[] static to
    arch/x86/kernel/apic/x2apic_uv_x.c.
In addition the initial touch_nmi_watchdog() before
the loop was removed and the time delay until it
is called within the loop was lengthened to 5 seconds.

This is UV only code and has been tested on a large
machine and previous similar patches on mulitple
machines.

Please ACK.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>