GFS2: rewrite fallocate code to write blocks directly
Patchwork-id: 42159
O-Subject: [RHEL 6.2 PATCH] GFS2: rewrite fallocate code to write blocks directly
Bugzilla: 750208
CVE: CVE-2011-4098
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Robert S Peterson <rpeterso@redhat.com>

The code to fix bz #737989 could cause a gfs2 process to try to start a new
transaction while already in a transaction.  To avoid that, this patch replaces
the original fix for #737989 with the upstream fix, which doesn't use the page
cache at all, since it's only writing to the end of the file or to holes in it.
Instead, it zeros the block out directly at allocation time.

This fixes bz #750208

Signed-off-by: Benjamin Marzinski <bmarzins@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>