netfilter: fix nf_conntrack refcount leak in l4proto->error()
Patchwork-id: 42185
O-Subject: [PATCH RHEL6.2] netfilter: fix nf_conntrack refcount leak in
	l4proto->error()
Bugzilla: 745472
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Jiri Pirko <jpirko@redhat.com>

BZ745472 describes a hang occurring occasionally while unloading the
nf_conntrack module. I have not been able to reproduce the issue
myself but based on the description it seems obvious that the module
hangs while waiting for all conntrack entries to die.

Upstream commit 88ed01d17b fixes a nfct leak when a l4 protocol error()
handler assign a conntrack to an skb. It is likely that we are seeing
this bug.

Since the issue is hard to reproduce we have not yet had the chance to
verify that the reported bug actually disappeared but this commit should
be backported regardless of this as a conntrack leak that can be triggered
by specially crafted packets is pretty serious.

Resolves BZ745472

Testing:
I have tested a kernel with the patch extensively to make sure this does
not break anything.

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=3766747

Upstream:
    commit 88ed01d17b44bc2bed4ad4835d3b1099bff3dd71
    Author: Pablo Neira Ayuso <pablo@netfilter.org>
    Date:   Thu Jun 2 15:08:45 2011 +0200

    netfilter: nf_conntrack: fix ct refcount leak in l4proto->error()

    This patch fixes a refcount leak of ct objects that may occur if
    l4proto->error() assigns one conntrack object to one skbuff. In
    that case, we have to skip further processing in nf_conntrack_in().

    With this patch, we can also fix wrong return values (-NF_ACCEPT)
    for special cases in ICMP[v6] that should not bump the invalid/error
    statistic counters.

    Reported-by: Zoltan Menyhart <Zoltan.Menyhart@bull.net>
    Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>

Signed-off-by: Thomas Graf <tgraf@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>