nv50/disp: shutdown display on suspend/hibernate
Patchwork-id: 42187
O-Subject: [RHEL6.2 PATCH] drm/nv50/disp: shutdown display on suspend/hibernate
Bugzilla: 740857
RH-Acked-by: Jerome Glisse <jglisse@redhat.com>
RH-Acked-by: Adam Jackson <ajax@redhat.com>

rhbz#740857

When using the NVIDIA GPU, the Dell M4600/M4600 systems currently fail
hibernation - either by spontaneously rebooting during the suspend process,
or by taking a very long time to go into hibernation and then failing to
restore the display on resume.

The solution to this problem is to shutdown the display engine completely
when freezing the GPU so the hibernation image can be prepared.

This patch is simplified to have minimal impact to the 6.2 kernel at this
late stage, and as such is not and will not be upstream in its current form.

More significant modifications will be performed upstream which will also
resolve this issue there.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>