fc class: fix building of Fibre Channel DUP drivers in 6.2
Patchwork-id: 42214
O-Subject: [RHEL6.2 PATCH 1/1] fc class: fix building of Fibre Channel DUP
	drivers in 6.2
Bugzilla: 750268
RH-Acked-by: Rob Evers <revers@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This for bz 750268.

The FC class code is *not* KABId, but this commit:

commit 0bdd4e1c305e4e2fad5c88bc995316e06c10efe1
Author: Mike Christie <mchristi@redhat.com>
Date:   Wed Oct 12 09:44:30 2011 -0400

    [scsi] scsi_transport_fc: Fix deadlock during fc_remove_host

added a new struct on the fc_rport which is breaking DUP drivers.

I guess this can be fixed by just adding __GENKSYMS__ around the
new field like in this patch. I am not sure if we want this patch though,
because the fc class code is just not KABId so the breakage is expected.
OTOH, it causes Hitachi a regression in their driver.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>