script: "sym" field really means show IP data
Patchwork-id: 42221
O-Subject: [PATCH RHEL6.3 BZ752353 03/65] perf script: "sym" field really means
	show IP data
Bugzilla: 752353
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 752353
https://bugzilla.redhat.com/show_bug.cgi?id=752353

upstream status
===============
commit 787bef174f055343c69a9639e6e05a564980ed4c
Author: David Ahern <dsahern@gmail.com>
Date:   Fri May 27 14:28:43 2011 -0600

description
===========
Currently the "sym" output field is used to dump instruction pointers
and callchain stack. Sample addresses can also be converted to symbols,
so the meaning of "sym" needs to be fixed. This patch adds an "ip"
option and if it is selected the user can also opt to dump symbols for
them. If the user opts to dump IP without syms only the address is
shown.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>