script: Make printing of dso a separate field option
Patchwork-id: 42222
O-Subject: [PATCH RHEL6.3 BZ752353 04/65] perf script: Make printing of dso a
	separate field option
Bugzilla: 752353
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 752353
https://bugzilla.redhat.com/show_bug.cgi?id=752353

upstream status
===============
commit 610723f24eeb842025178a6722fa9108c4e157b6
Author: David Ahern <dsahern@gmail.com>
Date:   Fri May 27 14:28:44 2011 -0600

description
===========
The 'sym' option displays both the function name and the DSO it comes
from. Split the display of the dso into a separate option.  This allows
display of the ip address and symbol without the dso, thus shortening
line lengths - and decluttering the output a bit.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>