tools: Allow sort dimensions to be registered more than once
Patchwork-id: 42228
O-Subject: [PATCH RHEL6.3 BZ752353 10/65] perf tools: Allow sort dimensions to
	be registered more than once
Bugzilla: 752353
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 752353
https://bugzilla.redhat.com/show_bug.cgi?id=752353

upstream status
===============
commit fd8ea21276adefc7f0133bd42fcf3b2faf0b15f8
Author: Frederic Weisbecker <fweisbec@gmail.com>
Date:   Wed Jun 29 23:08:14 2011 +0200

description
===========
So that the parent sort dimension can be registered twice: once
if we add it as an explicit sort dimension (-s parent) and twice
if we request a parent filter (-p foo).

We'll have only one parent sort dimension in the end but this
allows to override the default parent filter with we gave in "-p"
option. The goal of this is to prepare to allow the use of
"-s parent" and "-p foo" at the same time, ie: sort by filtered
parent.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>