tools: Fix error handling of unknown events
Patchwork-id: 42273
O-Subject: [PATCH RHEL6.3 BZ752353 48/65] perf tools: Fix error handling of
	unknown events
Bugzilla: 752353
RH-Acked-by: Don Zickus <dzickus@redhat.com>

Bugzilla: 752353
https://bugzilla.redhat.com/show_bug.cgi?id=752353

upstream status
===============
commit 777d1d71db622a5e1ff703495741c3d257b532e5
Author: Stephane Eranian <eranian@google.com>
Date:   Sat Jul 23 04:10:43 2011 +0200

description
===========
There was a problem with the parse_events() code not printing the
correct event name when an event was unknown and starting with an 'r'.
The source of the problem was the way raw notation was parsed.

Without the patch:
	$ perf stat -e retired_foo
	invalid event modifier: 'tired_foo'

With the patch:
	$ perf stat -e retired_foo
	invalid or unsupported event: 'retired_foo'

This also covers the case where the name of the event was not printed at
all when perf was linked with libpfm4.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>